# docs/swagger/swagger.json
{"openapi":"3.0.1","info":{"title":"XRT Training","version":"1.0.0"},"basePath":"/","components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"bearerAuth":[]}],"paths":{"/":{"get":{"description":"Default route","responses":{"200":{"description":"Success"}}}},"/auth":{"get":{"description":"Register a new user","tags":["Auth"],"responses":{"200":{"description":"Success"}}}},"/auth/register":{"post":{"description":"Register a new user","tags":["Auth"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","example":"test123"},"password":{"type":"string","example":"password123"},"firstname":{"type":"string","example":"john"},"lastname":{"type":"string","example":"smith"},"email":{"type":"string","example":"test@test.com"},"staffid":{"type":"string","example":"abc123"},"isSupervisor":{"type":"boolean","example":true}}}}}},"responses":{"200":{"description":"Success"}}}},"/auth/login":{"post":{"description":"Login","tags":["Auth"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","example":"sup1"},"password":{"type":"string","example":"test123"}}}}}},"responses":{"200":{"description":"Success"}}}},"/course/progress":{"post":{"description":"Post a progress update","tags":["Progress"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","example":{"any":"thing"}}}}}}},"responses":{"200":{"description":"Success"}}},"get":{"description":"Get all progress updates","tags":["Progress"],"responses":{"200":{"description":"Success"}}}},"/course":{"get":{"description":"Get all the courses","tags":["Course"],"responses":{"200":{"description":"Success"}}},"post":{"description":"Get a course by id","tags":["Course"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Course"},"description":{"type":"string","example":"This is the first of many courses"},"tasks":{"type":"array","items":{"type":"string","example":"task01"},"example":["task01","task02"]},"assignedEmployees":{"type":"array","items":{"type":"string","example":"user01"},"example":["user01","user02"]},"dueDate":{"type":"Date","example":"05/09/2021"}}}}}},"responses":{"200":{"description":"Success"}}}},"/course/{courseId}":{"get":{"description":"Get a course by id","tags":["Course"],"parameters":[{"in":"path","name":"courseId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Update a course by id","tags":["Course"],"parameters":[{"in":"path","name":"courseId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Course"},"description":{"type":"string","example":"This is the first of many courses"},"tasks":{"type":"array","items":{"type":"string","example":"task01"},"example":["task01","task02"]},"assignedEmployees":{"type":"array","items":{"type":"string","example":"user01"},"example":["user01","user02"]},"dueDate":{"type":"Date","example":"yyyy-mm-dd"}}}}}},"responses":{"200":{"description":"Success"}}},"delete":{"description":"Get a course by id","tags":["Course"],"parameters":[{"in":"path","name":"courseId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/floor":{"get":{"description":"Get all floors","tags":["Floor"],"responses":{"200":{"description":"Success"}}},"post":{"description":"Create a floor","tags":["Floor"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Floor"},"data":{"type":"object","example":"data"}}}}}},"responses":{"200":{"description":"Success"}}}},"/floor/{floorId}":{"get":{"description":"Get a floor by id","tags":["Floor"],"parameters":[{"in":"path","name":"floorId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Update a floor by id","tags":["Floor"],"parameters":[{"in":"path","name":"floorId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Floor"},"data":{"type":"object","example":"data"}}}}}},"responses":{"200":{"description":"Success"}}},"delete":{"description":"Delete a floor","tags":["Floor"],"parameters":[{"in":"path","name":"floorId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/upload":{"post":{"description":"Upload an Image","tags":["User"],"requestBody":{"required":"true,","content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"Success"}}}},"/upload/{username}":{"get":{"description":"Get profile image for user","tags":["User"],"parameters":[{"in":"path","name":"username","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/ingredient":{"get":{"description":"Get all ingredients","tags":["Ingredient"],"responses":{"200":{"description":"Success"}}},"post":{"description":"Create an ingredient","tags":["Ingredient"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"Tomato"},"id":{"type":"string","example":"VEG1"}}}}}},"responses":{"200":{"description":"Success"}}}},"/ingredient/search":{"get":{"description":"Get an ingredient by id","tags":["Ingredient"],"parameters":[{"in":"query","name":"id","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/progress":{"put":{"description":"Create or Updates a progress","tags":["Progress"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"Object","example":{"data":"data"}},"userId":{"type":"string","example":"John"},"taskId":{"type":"string","example":"Task"},"courseId":{"type":"string","example":"Course"},"completed":{"type":"Boolean","example":true},"score":{"type":"Number","example":22}}}}}},"responses":{"200":{"description":"Success"}}},"get":{"description":"Get a progress by userId, taskId, courseId or all of them","tags":["Progress"],"parameters":[{"in":"query","name":"userId","required":false,"type":"string"},{"in":"query","name":"taskId","required":false,"type":"string"},{"in":"query","name":"courseId","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/progress/statistics":{"get":{"description":"Get a progress by userId, taskId, courseId or all of them","tags":["Progress"],"parameters":[{"in":"query","name":"userId","required":false,"type":"string"},{"in":"query","name":"taskId","required":false,"type":"string"},{"in":"query","name":"courseId","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/progress/tracking":{"put":{"description":"Put tracking logs on a progress by userId, taskId and courseId","tags":["Progress"],"parameters":[{"in":"query","name":"userId","required":false,"type":"string"},{"in":"query","name":"taskId","required":false,"type":"string"},{"in":"query","name":"courseId","required":false,"type":"string"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"date":{"type":"Date","example":"2021-09-12T14:05:47.823Z"},"event":{"type":"string","example":"FRIDGE"},"value":{"type":"string","example":"OPEN"},"data":{"type":"Object","example":{"data":"data"}}}}}}},"responses":{"200":{"description":"Success"}}}},"/progress/tracking/logs":{"get":{"description":"Get tracking logs on a progress for a userId and courseId","tags":["Progress"],"parameters":[{"in":"query","name":"userId","required":false,"type":"string"},{"in":"query","name":"courseId","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/recipe/category":{"get":{"description":"Get all the Categoriesnnp","tags":["Recipe"],"responses":{"200":{"description":"Success"}}}},"/recipe":{"get":{"description":"Get all the recipes","tags":["Recipe"],"responses":{"200":{"description":"Success"}}},"post":{"description":"Create a recipe","tags":["Recipe"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"Whooper"},"steps":{"type":"array","items":{"type":"object","example":["1) Put bottom bun on plate"]},"example":["1) Put bottom bun on plate","2) Insert beef patty","3) Place lettuce on top","4) Put on top bun"]},"ingredients":{"type":"array","items":{"type":"string","example":"top_bun"},"example":["top_bun","cheese","patty","bottom_bun"]},"category":{"type":"string","example":"Cheeseburger"}}}}}},"responses":{"200":{"description":"Success"}}}},"/recipe/search":{"get":{"description":"Get recipes by category","tags":["Recipe"],"parameters":[{"in":"query","name":"category","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/recipe/{recipeId}":{"get":{"description":"Get a recipe by id","tags":["Recipe"],"parameters":[{"in":"path","name":"recipeId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Update a recipe by id","tags":["Recipe"],"parameters":[{"in":"path","name":"recipeId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"Whooper"},"steps":{"type":"array","items":{"type":"object","example":["1) Put bottom bun on plate"]},"example":["1) Put bottom bun on plate","2) Insert beef patty","3) Place lettuce on top","4) Put on top bun"]},"ingredients":{"type":"array","items":{"type":"string","example":"top_bun"},"example":["top_bun","cheese","patty","bottom_bun"]},"category":{"type":"string","example":"Cheeseburger"}}}}}},"responses":{"200":{"description":"Success"}}},"delete":{"description":"Delete a recipe","tags":["Recipe"],"parameters":[{"in":"path","name":"recipeId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/task":{"get":{"description":"Get all tasks","tags":["Task"],"responses":{"200":{"description":"Success"}}},"post":{"description":"Create a task","tags":["Task"],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Task"},"description":{"type":"string","example":"Boil water"},"recipe":{"type":"string","example":{"name":"Empty Burger","steps":[],"ingredients":["top_bun","bottom_bun"],"category":""}},"type":{"type":"string","example":"liquid"}}}}}},"responses":{"200":{"description":"Success"}}}},"/task/{taskId}":{"get":{"description":"Get all tasks by name","tags":["Task"],"parameters":[{"in":"path","name":"taskId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Update a task by id","tags":["Task"],"parameters":[{"in":"path","name":"taskId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","example":"First Task"},"description":{"type":"string","example":"Boil water"},"type":{"type":"string","example":"liquid"},"recipe":{"type":"string","example":"recipe01"}}}}}},"responses":{"200":{"description":"Success"}}},"delete":{"description":"Delete a task","tags":["Task"],"parameters":[{"in":"path","name":"taskId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/task/getById/{userId}":{"get":{"description":"Get all tasks assigned to a certain Id","tags":["Task"],"parameters":[{"in":"path","name":"userId","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/task/getById/not/{userId}":{"get":{"description":"Get all tasks not assigned to a certain Id","tags":["Task"],"parameters":[{"in":"path","name":"userId","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/user/all":{"get":{"description":"Get all the users","tags":["User"],"responses":{"200":{"description":"Success"}}}},"/user":{"get":{"description":"Get current user","tags":["User"],"responses":{"200":{"description":"Success"}}}},"/user/search":{"get":{"description":"Get all supervisors or Get all trainees","tags":["User"],"parameters":[{"in":"query","name":"isSupervisor","required":true,"type":"boolean"}],"responses":{"200":{"description":"Success"}}}},"/user/{userId}":{"get":{"description":"Get a user by id","tags":["User"],"parameters":[{"in":"path","name":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}},"put":{"description":"Get a user by id","tags":["User"],"parameters":[{"in":"path","name":"userId","required":true,"type":"string"}],"requestBody":{"required":"true,","content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","example":"test123"},"password":{"type":"string","example":"password123"},"newPassword":{"type":"string","example":"password123"},"firstname":{"type":"string","example":"john"},"lastname":{"type":"string","example":"smith"},"email":{"type":"string","example":"test@test.com"},"staffid":{"type":"string","example":"abc123"},"isSupervisor":{"type":"boolean","example":true}}}}}},"responses":{"200":{"description":"Success"}}},"delete":{"description":"Delete a task","tags":["User"],"parameters":[{"in":"path","name":"userId","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}}},"tags":[]}